Remove WindDataDisplay component and add SodaLakeScreen for wind data visualization and analysis
